baseline/requirements.txt
-i https://pypi.org/simple
datasets==3.2.0
huggingface-hub==0.27.1
smolagents==1.3.0
litellm==1.58.2
tenacity==9.0.0
azure-identity>=1.12.0
azure-core>=1.24.0
openai>=1.12.0

